Create Hide-Save-and-New-on-Account-Edit-HTML-Code
Save And New Button Appears on the Account Edit Screen Online - 
Veeva New Account Wizards provide a way to ensure that all new accounts created by users also include an address.  We found that editing an Account record would allow creation of an Account WITHOUT an Address record due to the display of the Salesforce “Save and New” button.    We need to hide the Save and New Button on Account Edit.